chore: rename UserDetailsTooltip.js to index.js
